refactor: remove cycle from Taxes
